requirements.txt
biopython==1.85
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-CORS==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
scipy==1.11.3
Werkzeug==3.1.3
